+++ skrobot/model/link.py
@@ -1,0 +1,118 @@
+import collections
+
+import numpy as np
+import trimesh
+
+from skrobot.coordinates import CascadedCoords
+
+
+class Link(CascadedCoords):
+
+    def __init__(self, centroid=None,
+                 inertia_tensor=None,
+                 collision_mesh=None,
+                 visual_mesh=None,
+                 *args, **kwargs):
+        super(Link, self).__init__(*args, **kwargs)
+        self.centroid = centroid
+        self.joint = None
+        self._child_links = []
+        self._parent_link = None
+        if inertia_tensor is None:
+            inertia_tensor = np.eye(3)
+        self._collision_mesh = collision_mesh
+        self._visual_mesh = visual_mesh
+
+    @property
+    def parent_link(self):
+        return self._parent_link
+
+    @property
+    def child_links(self):
+        return self._child_links
+
+    def add_joint(self, j):
+        self.joint = j
+
+    def delete_joint(self):
+        self.joint = None
+
+    def add_child_link(self, child_link):
+        """Add child link."""
+        if child_link is not None and child_link not in self._child_links:
+            self._child_links.append(child_link)
+
+    def del_child_link(self, link):
+        self._child_links.remove(link)
+
+    def add_parent_link(self, parent_link):
+        self._parent_link = parent_link
+
+    def del_parent_link(self):
+        self._parent_link = None
+
+    @property
+    def collision_mesh(self):
+        """Return collision mesh
+
+        Returns
+        -------
+        self._collision_mesh : trimesh.base.Trimesh
+            A single collision mesh for the link.
+            specified in the link frame,
+            or None if there is not one.
+        """
+        return self._collision_mesh
+
+    @collision_mesh.setter
+    def collision_mesh(self, mesh):
+        """Setter of collision mesh
+
+        Parameters
+        ----------
+        mesh : trimesh.base.Trimesh
+            A single collision mesh for the link.
+            specified in the link frame,
+            or None if there is not one.
+        """
+        if mesh is not None and \
+           not isinstance(mesh, trimesh.base.Trimesh):
+            raise TypeError('input mesh is should be trimesh.base.Trimesh, '
+                            'get type {}'.format(type(mesh)))
+        self._collision_mesh = mesh
+
+    @property
+    def visual_mesh(self):
+        """Return visual mesh
+
+        Returns
+        -------
+        self._visual_mesh : None, trimesh.base.Trimesh, or
+                            sequence of trimesh.Trimesh
+            A set of visual meshes for the link in the link frame.
+        """
+        return self._visual_mesh
+
+    @visual_mesh.setter
+    def visual_mesh(self, mesh):
+        """Setter of visual mesh
+
+        Parameters
+        ----------
+        mesh : None, trimesh.Trimesh, sequence of trimesh.Trimesh,
+               trimesh.points.PointCloud or str
+            A set of visual meshes for the link in the link frame.
+        """
+        if not (mesh is None
+                or isinstance(mesh, trimesh.Trimesh)
+                or (isinstance(mesh, collections.Sequence)
+                    and all(isinstance(m, trimesh.Trimesh) for m in mesh))
+                or isinstance(mesh, trimesh.points.PointCloud)
+                or isinstance(mesh, str)):
+            raise TypeError(
+                'mesh must be None, trimesh.Trimesh, sequence of '
+                'trimesh.Trimesh, trimesh.points.PointCloud '
+                'or path of mesh file, but got: {}'.format(type(mesh)))
+        if isinstance(mesh, str):
+            mesh = trimesh.load(mesh)
+        self._visual_mesh = mesh
