+++ skrobot/models/__init__.py
@@ -1,15 +1,5 @@
 # flake8: noqa
 
-from skrobot.models.primitives import Annulus
-from skrobot.models.primitives import Axis
-from skrobot.models.primitives import Box
-from skrobot.models.primitives import CameraMarker
-from skrobot.models.primitives import Cone
-from skrobot.models.primitives import Cylinder
-from skrobot.models.primitives import MeshLink
-from skrobot.models.primitives import PointCloudLink
-from skrobot.models.primitives import Sphere
-
 from skrobot.models.fetch import Fetch
 from skrobot.models.kuka import Kuka
 from skrobot.models.panda import Panda
